refactor(npm): Split packager into a transformer and put result in new package.
