+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/npm/cli/NpmCliExtractor.java
@@ -20,7 +20,7 @@ import com.synopsys.integration.detectable.ExecutableUtils;
 import com.synopsys.integration.detectable.detectable.executable.DetectableExecutableRunner;
 import com.synopsys.integration.detectable.detectables.npm.cli.parse.NpmCliParser;
 import com.synopsys.integration.detectable.detectables.npm.cli.parse.NpmDependencyTypeFilter;
-import com.synopsys.integration.detectable.detectables.npm.lockfile.model.NpmParseResult;
+import com.synopsys.integration.detectable.detectables.npm.lockfile.result.NpmPackagerResult;
 import com.synopsys.integration.detectable.detectables.npm.packagejson.model.PackageJson;
 import com.synopsys.integration.detectable.extraction.Extraction;
 import com.synopsys.integration.detectable.util.ToolVersionLogger;
@@ -76,7 +76,7 @@ public class NpmCliExtractor {
             logger.debug("Parsing npm ls file.");
             logger.debug(standardOutput);
             NpmDependencyTypeFilter npmDependencyTypeFilter = new NpmDependencyTypeFilter(packageJson.devDependencies.keySet(), packageJson.peerDependencies.keySet(), includeDevDependencies, includePeerDependencies);
-            NpmParseResult result = npmCliParser.generateCodeLocation(standardOutput, npmDependencyTypeFilter);
+            NpmPackagerResult result = npmCliParser.generateCodeLocation(standardOutput, npmDependencyTypeFilter);
             String projectName = result.getProjectName() != null ? result.getProjectName() : packageJson.name;
             String projectVersion = result.getProjectVersion() != null ? result.getProjectVersion() : packageJson.version;
             return new Extraction.Builder().success(result.getCodeLocation()).projectName(projectName).projectVersion(projectVersion).build();
